+++ src/machine-c-api.h
@@ -360,6 +360,7 @@ typedef struct { // NOLINT(modernize-use-using)
 /// \brief Machine runtime configuration
 typedef struct { // NOLINT(modernize-use-using)
     cm_concurrency_config concurrency;
+    bool skip_root_hash_check;
 } cm_machine_runtime_config;
 
 /// \brief Machine instance handle